requirements.txt
pyrogram==2.0.96
tgcrypto==1.2.5
httpx
pillow
cashews
coloredlogs
qrcode
pyncm
mutagen
playwright
uvicorn
jinja2
apscheduler
pytz
python-twitter
beautifulsoup4
lxml
sqlalchemy
sqlmodel
aiosqlite
pydantic
